refactor(netlist): isolate the deprecated old renderer
render_netlist -- the original gamma/mu/eta/`:=` text form, superseded by
render_lvnet and kept only as a parity reference until Phase E retires it --
moves to a new netlist_render_deprecated.py (432 lines) with its exclusive
helpers. ambiguous_bares / instance_line / scope_header are shared with
diff.py, so they move alongside and netlist.py re-exports them (diff's imports
stay unchanged). Named "deprecated", not "legacy": it's a superseded renderer,
NOT format back-compat. netlist.py 3,717 -> 3,372.

Byte-identical lvnet; pyright src/ clean; the netlist/lvnet/parity test subset
green; one-way dependency, no import cycle.
